[ios] Move most of iOS code to new apple_embedded driver
No meaningful code changes other than file renames.
