+++ drivers/apple_embedded/device_metrics.mm
