+++ drivers/apple_embedded/display_layer.mm
@@ -30,8 +30,8 @@
 
 #import "display_layer.h"
 
-#import "display_server_ios.h"
-#import "os_ios.h"
+#import "display_server.h"
+#import "os_apple_embedded.h"
 
 #include "core/config/project_settings.h"
 #include "core/os/keyboard.h"
