+++ drivers/apple_embedded/display_server.h
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  display_server_ios.h                                                  */
+/*  display_server.h                                                      */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             GODOT ENGINE                               */
@@ -38,7 +38,7 @@
 #include "servers/rendering/rendering_device.h"
 
 #if defined(VULKAN_ENABLED)
-#import "rendering_context_driver_vulkan_ios.h"
+#import "rendering_context_driver_vulkan.h"
 
 #include "drivers/vulkan/godot_vulkan.h"
 #endif // VULKAN_ENABLED
