+++ drivers/apple_embedded/godot_app_delegate.mm
@@ -30,7 +30,7 @@
 
 #import "godot_app_delegate.h"
 
-#import "app_delegate.h"
+#import "app_delegate_service.h"
 
 @interface GodotApplicationDelegate ()
