+++ platform/ios/SCsub
@@ -60,23 +60,7 @@ def generate_bundle(target, source, env):
 
 
 ios_lib = [
-    "godot_ios.mm",
-    "os_ios.mm",
     "main.m",
-    "app_delegate.mm",
-    "view_controller.mm",
-    "ios.mm",
-    "rendering_context_driver_vulkan_ios.mm",
-    "display_server_ios.mm",
-    "godot_view.mm",
-    "tts_ios.mm",
-    "display_layer.mm",
-    "godot_app_delegate.m",
-    "godot_view_renderer.mm",
-    "device_metrics.m",
-    "keyboard_input_view.mm",
-    "key_mapping_ios.mm",
-    "ios_terminal_logger.mm",
 ]
 
 env_ios = env.Clone()
